Create MaxSum
You are given an integer array arr, consisting of N integers. Find the maximum possible sum of consecutive numbers in arr. For example: if the array arr consists of the numbers 1, 6, -9, 4, 4, -2, 10, -1, the maximum possible sum of consecutive numbers is 16 (the consecutive numbers are 4, 4, -2 and 10) .Element 0 of the array is the number N. Next N elements (from 1 to N) construct the array arr.Your method should return a single number - the maximum possible sum of consecutive numbers.
Example input	Example output
8
1
6
-9
4
4
-2
10
-1	16

Example input	Example output
6
1
3
-5
8
7
-6	15